+++ web/documentserver-example/csharp/DocEditor.aspx.cs
@@ -207,6 +207,17 @@ namespace OnlineEditorsExample
                                         }
                                 },
                                 {
+                                    "referenceData", new Dictionary<string, string>()
+                                    {
+                                        { "fileKey", !user.id.Equals("uid-0") ?
+                                            jss.Serialize(new Dictionary<string, object>{
+                                                {"fileName", FileName},
+                                                {"userAddress", HttpUtility.UrlEncode(_Default.CurUserHostAddress(HttpContext.Current.Request.UserHostAddress))}
+                                        }) : null },
+                                        {"instanceId", _Default.GetServerUrl(false) }
+                                    }
+                                },
+                                {
                                     // the permission for the document to be edited and downloaded or not
                                     "permissions", new Dictionary<string, object>
                                         {
@@ -222,7 +233,8 @@ namespace OnlineEditorsExample
                                             { "chat", !user.id.Equals("uid-0") },
                                             { "reviewGroups", user.reviewGroups },
                                             { "commentGroups", user.commentGroups },
-                                            { "userInfoGroups", user.userInfoGroups }
+                                            { "userInfoGroups", user.userInfoGroups },
+                                            { "protect", !user.deniedPermissions.Contains("protect") }
                                         }
                                 }
                             }