+++ web/documentserver-example/csharp/Users.cs
@@ -68,6 +68,7 @@ namespace OnlineEditorsExample
             "Can’t see anyone’s information",
             "Can't rename files from the editor",
             "Can't view chat",
+            "Can't protect file",
             "View file without collaboration",
         };
 
@@ -130,7 +131,7 @@ namespace OnlineEditorsExample
                     new Dictionary<string, object>(),
                     new List<string>(),
                     null,
-                    new List<string>(),
+                    new List<string>() { "protect" },
                     descr_user_0,
                     false
                 )