+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/documentserver/models/filemodel/Document.java
@@ -18,12 +18,15 @@
 
 package com.onlyoffice.integration.documentserver.models.filemodel;
 
+import com.onlyoffice.integration.documentserver.managers.document.DocumentManager;
 import com.onlyoffice.integration.documentserver.models.configurations.Info;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
+
+import java.util.HashMap;
 
 @Component
 @Scope("prototype")
@@ -43,4 +46,5 @@ public class Document {  // the parameters pertaining to the document (title, ur
     as file name when the document is downloaded */
     private String url;  // the absolute URL where the source viewed or edited document is stored
     private String directUrl;
+    private HashMap<String, Object> referenceData;
 }