+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/entities/Permission.java
@@ -52,4 +52,5 @@ public class Permission extends AbstractEntity {
     private List<Group> commentsRemoveGroups;
     @ManyToMany
     private List<Group> userInfoGroups;
+    private Boolean protect = true;
 }