+++ web/documentserver-example/nodejs/public/javascripts/jscript.js
@@ -245,7 +245,7 @@ if (typeof jQuery != "undefined") {
         window.open(url, "_blank");
         jq('#hiddenFileName').val("");
         jq.unblockUI();
-        document.location.reload();
+        document.location.reload(true);
     });
 
     jq(document).on("click", "#beginView:not(.disable)", function () {
@@ -258,7 +258,7 @@ if (typeof jQuery != "undefined") {
         window.open(url, "_blank");
         jq('#hiddenFileName').val("");
         jq.unblockUI();
-        document.location.reload();
+        document.location.reload(true);
     });
 
     jq(document).on("click", "#beginEmbedded:not(.disable)", function () {
@@ -272,13 +272,13 @@ if (typeof jQuery != "undefined") {
     });
 
     jq(document).on("click", ".reload-page", function () {
-        setTimeout(function () { document.location.reload(); }, 1000);
+        setTimeout(function () { document.location.reload(true); }, 1000);
         return true;
     });
 
     jq(document).on("mouseup", ".reload-page", function (event) {
         if (event.which == 2) {
-            setTimeout(function () { document.location.reload(); }, 1000);
+            setTimeout(function () { document.location.reload(true); }, 1000);
         }
         return true;
     });
@@ -288,7 +288,7 @@ if (typeof jQuery != "undefined") {
         jq("#embeddedView").remove();
         jq.unblockUI();
         if (mustReload) {
-            document.location.reload();
+            document.location.reload(true);
         }
     });
 
@@ -303,7 +303,7 @@ if (typeof jQuery != "undefined") {
             type: "delete",
             url: requestAddress,
             complete: function (data) {
-                document.location.reload();
+                document.location.reload(true);
             }
         });
     });