+++ web/documentserver-example/php/doceditor.php
@@ -18,7 +18,6 @@
 namespace OnlineEditorsExamplePhp;
 
 use OnlineEditorsExamplePhp\Views\DocEditorView;
-
 require_once dirname(__FILE__) . '/functions.php';
 require_once dirname(__FILE__) . '/vendor/autoload.php';