+++ web/documentserver-example/ruby/app/controllers/home_controller.rb
@@ -363,4 +363,52 @@ class HomeController < ApplicationController
       json_data = TrackHelper.command_request("meta", dockey, meta)
       render plain: '{ "result" : "' + JSON.dump(json_data) + '"}'
     end
+
+    #ReferenceData
+    def reference
+      body = JSON.parse(request.body.read)
+      fileName = ""
+      
+
+      if body.key?("referenceData")
+        referenceData = body["referenceData"]
+        instanceId = referenceData["instanceId"]
+        if instanceId == DocumentHelper.get_server_url(false)
+          fileKey = JSON.parse(referenceData["fileKey"])
+          userAddress = fileKey["userAddress"]
+          if userAddress == DocumentHelper.cur_user_host_address(nil)
+            fileName = fileKey["fileName"]
+          end
+        end
+      end
+
+      if fileName.empty? and body.key?("path")
+        path = File.basename(body["path"])
+        if File.exist?(DocumentHelper.storage_path(path, nil))
+          fileName = path
+        end
+      end
+
+      if fileName.empty?
+        render plain: '{ "error": "File not found"}'
+        return
+      end
+
+      data = {
+        :fileType => DocumentHelper.get_internal_extension(fileName),
+        :url => DocumentHelper.get_download_url(fileName),
+        :directUrl => body["directUrl"] ? DocumentHelper.get_download_url(fileName) : DocumentHelper.get_download_url(fileName,false),
+        :referenceData => {
+          :instanceId => DocumentHelper.get_server_url(false),
+          :fileKey => {:fileName => fileName,:userAddress => DocumentHelper.cur_user_host_address(nil)}.to_json
+        },
+        :path => fileName
+      }
+
+      if JwtHelper.is_enabled
+        data["token"] = JwtHelper.encode(data)
+      end
+
+      render plain: data.to_json
+    end
 end